+++ dbrepo-ui/pages/database/[database_id]/subset/[subset_id]/data.vue
@@ -36,7 +36,8 @@
     </v-toolbar>
     <v-card
       v-if="subset"
-      tile>
+      rounded="0"
+      variant="flat">
       <QueryResults
         id="query-results"
         ref="queryResults"